refactor tool operation
