+++ src/promptflow/promptflow/_sdk/_constants.py
@@ -125,14 +125,6 @@ REGISTRY_URI_PREFIX = "azureml://registries/"
 FLOW_RESOURCE_ID_PREFIX = "azureml://locations/"
 FLOW_DIRECTORY_MACRO_IN_CONFIG = "${flow_directory}"
 
-# Tool meta info
-UIONLY_HIDDEN = "uionly_hidden"
-SKIP_FUNC_PARAMS = ["subscription_id", "resource_group_name", "workspace_name"]
-ICON_DARK = "icon_dark"
-ICON_LIGHT = "icon_light"
-ICON = "icon"
-TOOL_SCHEMA = Path(__file__).parent / "data" / "tool.schema.json"
-
 # trace
 TRACE_MGMT_DB_PATH = (HOME_PROMPT_FLOW_DIR / "trace.sqlite").resolve()
 TRACE_MGMT_DB_SESSION_ACQUIRE_LOCK_PATH = (HOME_PROMPT_FLOW_DIR / "trace.sqlite.lock").resolve()
